+++ CHANGELOG.md
@@ -1,11 +1,9 @@
-# Version 2.2.1-alpha
-- Initial power framework support.
-
 # Version 2.2.0-alpha
 - Defensive powers are used directly.  No longer need to have separate defense items splitting out PD/ED/etc.
 - Reworked ActiveEffects to be placed on items (per FoundryVtt design).
 - Apply damage only shown to GMs [#95](https://github.com/dmdorman/hero6e-foundryvtt/issues/95)
 - Power/item descriptions can be sent to chat [#128](https://github.com/dmdorman/hero6e-foundryvtt/issues/128)
+- Initial power framework support.
 
 # Version 2.1.9-alpha
 - Fixed equipment price showing NaN.  Summary weight/price for equipment now only shows when there are items with weight/price.